Changes to build for switch to Qt 6.6.3 64 bits + small win/qt fixes

+++ src/utils/pathut.h
@@ -115,8 +115,8 @@ bool path_readable(const std::string& path);
 
 // Conversion between utf-8 and wide char file names.
 
-bool wchartoutf8(const wchar_t *in, std::string& out, size_t len = 0);
-std::string wchartoutf8(const wchar_t *in, size_t len = 0);
+bool wchartoutf8(const wchar_t *in, std::string& out, int len = 0);
+std::string wchartoutf8(const wchar_t *in, int len = 0);
 bool utf8towchar(const std::string& in, wchar_t *out, size_t obytescap);
 std::unique_ptr<wchar_t[]> utf8towchar(const std::string& in);
